requirements.txt
streamlit
google-genai
plotly
pydantic